fix(webcomponents): move shortMimeLabel out of the global screenshot area
The `screenshots` job was cancelled on all three revisions of this PR. Not
runner congestion, as I first assumed: it ran 30m30s against a
`timeout-minutes: 30` ceiling and died in the R2 upload step, because it
was uploading 952 PNGs.

`storybook-affected-stories-lib.mjs` treats `src/internal/` as a GLOBAL
AREA — a change there fans out to EVERY story in the library — and the
comment says why: that directory holds the shared DOM and icon primitives
which really do change how every component renders. `shortMimeLabel` is a
pure string helper that alters two chips. Filing it there was wrong about
what the directory means, and the resolver was right to believe me.

Moved beside its in-library consumer as `quick-look/mime-label.ts`; the
subpath export follows it. Measured against the real story index with the
PR's actual changed-file list: 475 affected stories before, 224 after.

The remaining 224 is the honest floor and not worth engineering away —
both chat components changed (116 stories in that area) and a new
`primitives/` component was added (97). Those are the per-area rule
working as designed on changes that really did happen.
