+++ docs/webapp-details.md
@@ -117,7 +117,7 @@ One roster, three renderings, one vocabulary — `ui/follower-presentation.ts` o
 A payload PASTED into chat — a `data:` URL, the output of `base64 < key.pem` — collapses to a `<slicc-blob-chip>` that opens it in Quick Look. Same four roles as file mentions (`packages/webapp/CLAUDE.md`), and reusing the same machinery rather than growing a second copy of it.
 
 - Path: `core/base64-mentions.ts` (heuristic, pure) → `core/base64-payload.ts` (decode + identify) → `ui/base64-preview-linker.ts` (DOM swap) → `ui/wc/wire-base64-previews.ts` (lifecycle + Quick Look, wired from `mountWcShell`). Chip: `<slicc-blob-chip>` in `@slicc/webcomponents`.
-- **Consolidated, not duplicated.** The base64 grammar is `normalizeBase64` in `@slicc/shared-ts` (shared with `base/image-markers.ts`); decoding is `base64ToUint8`; identification composes `sniffMagicBytes` / `looksLikeText` / `isTextMimeType` from `core/file-type.ts`; the synthetic filename comes from `extensionForMimeType` in `base/mime-types.ts` (reverse index of the SAME table `getMimeType` reads); the rendered view is `file-actions.ts`'s `buildRenderedView`, so the 512 KB cap and the inline-vs-sandbox rule have one home; the chip's short type label is `@slicc/webcomponents/internal/mime-label`, shared with Quick Look's header chip.
+- **Consolidated, not duplicated.** The base64 grammar is `normalizeBase64` in `@slicc/shared-ts` (shared with `base/image-markers.ts`); decoding is `base64ToUint8`; identification composes `sniffMagicBytes` / `looksLikeText` / `isTextMimeType` from `core/file-type.ts`; the synthetic filename comes from `extensionForMimeType` in `base/mime-types.ts` (reverse index of the SAME table `getMimeType` reads); the rendered view is `file-actions.ts`'s `buildRenderedView`, so the 512 KB cap and the inline-vs-sandbox rule have one home; the chip's short type label is `@slicc/webcomponents/quick-look/mime-label`, shared with Quick Look's header chip.
 - **Three shapes, not one.** A `data:` URL, a bare single-line run, and the COLUMN-WRAPPED block `base64`(1) emits by default (76 columns; PEM uses 64). The wrapped shape is the advertised `base64 < report.pdf` scenario and its individual lines are each far below the length bar, so it needs its own pass — believed only when it has the exact shape an encoder produces: two or more whole lines, every one but the last the same width and a whole number of quanta, only the last carrying padding, and no more of the block after that last line (without which a ragged 76-then-72 pair of PROSE lines reads as a complete block). A first line that carries the user's own words in front of the paste (`here it is: <76 chars>`) is glued back on when its trailing run is exactly one wrap column wide — otherwise the block would be claimed from its second line, stranding the first 76 characters as text beside the chip.
 - **The DOM half has to match.** `marked` runs with `breaks: true`, so a wrapped paste arrives as `text <br> text <br> text` — one node per line. `ui/base64-preview-linker.ts` therefore walks SEGMENTS (maximal runs of adjacent `Text`/`<br>` siblings, joined with `\n` per `<br>`), not individual text nodes; any other element ends the run, because across a `<strong>` the text was never contiguous. `<br>`s that fall inside an elided span are removed with it, or a collapsed block leaves a stack of blank lines where its wrapping used to be.
 - **Eliding is destructive in a way linking is not.** A wrong file mention is a dead link; a wrong base64 match HIDES text the user typed. So the heuristic is the conservative twin of `file-mentions.ts`: ≥128 alphabet characters (`MIN_PAYLOAD_CHARS`), a whole number of quanta, and an explicit small delimiter set — a leading `.` opens nothing, so a JWT's middle segment is never clipped out from between its neighbours, while a trailing `.` closes a run so a payload can end a sentence. Then the bytes still have to decode to something recognizable: `identifyBase64` returns `null` (magic → declared → UTF-8 decodability, all misses) for a hash or an opaque id, and the run stays exactly as typed.
