+++ packages/webcomponents/src/quick-look/mime-label.ts
@@ -6,13 +6,20 @@
  * kind of thing it is showing (an image, a document) only needs the last
  * distinguishing word, and the full type is what the `title` is for.
  *
- * Lives in `internal/` rather than beside its first caller because two
+ * Its own module, beside Quick Look rather than inside it, because two
  * surfaces label the same payload and must agree: Quick Look's header chip,
  * and the transcript's decoded-base64 chip in the webapp (which reaches this
- * through the `@slicc/webcomponents/internal/*` entry, the way
- * `message-renderer.ts` already reaches `escapeHtml`). Two copies of this
- * would let a blob be called one thing in the bubble and another in the
- * previewer it opens.
+ * through the `@slicc/webcomponents/quick-look/mime-label` entry, the way
+ * `message-renderer.ts` reaches `escapeHtml`). Two copies would let a blob be
+ * called one thing in the bubble and another in the previewer it opens.
+ *
+ * Pointedly NOT in `internal/`, which is where it started. That directory is a
+ * GLOBAL AREA to the Storybook screenshot resolver
+ * (`storybook-affected-stories-lib.mjs`): it holds the shared DOM and icon
+ * primitives that really do change how every component renders, so a change
+ * there fans out to every story in the library. This is a pure string helper
+ * that alters two chips, and filing it there made a one-line refactor request
+ * 952 screenshots and time the upload job out.
  */
 
 /**
